% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feldkamp.R
\docType{data}
\name{feldkamp}
\alias{feldkamp}
\title{Feldkamp}
\format{
\subsection{\code{feldkamp} A data frame with 5,429 rows and 3 columns:}{

\describe{
\item{id}{Participant ID.}
\item{case}{Case of gastroschisis in newborn.}
\item{m_inf}{Genitourinary infection in the month before conception to the end of the first trimester, in the mother.}
}
}
}
\source{
\url{https://canvas.education.lu.se/courses/34580/files/6061948?module_item_id=1441713}
}
\usage{
feldkamp
}
\description{
A dataset used in \href{https://canvas.education.lu.se/courses/34580/files/6116314?module_item_id=1441704}{exercise 1}.
}
\keyword{datasets}
