% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodpressure.R
\docType{data}
\name{bloodpressure}
\alias{bloodpressure}
\title{Blood Pressure}
\format{
\subsection{\code{bloodpressure} A data frame with 429 rows and 11 columns:}{

\describe{
\item{id}{Participant ID.}
\item{sex}{Sex.}
\item{age}{Age.}
\item{BMI}{Body mass index.}
\item{SCL}{Serum cholesterol.}
\item{SBP}{Systolic variable.}
\item{DBP}{Diastolic variable.}
\item{ht}{Hypertension. TRUE if SBP ≥ 140 or DBP ≥ 90, FALSE otherwise.}
\item{chd_event}{Coronary heart disease event at last follow-up.}
\item{smoker}{Smoker.}
\item{fu_years}{Follow-up years.}
}
}
}
\source{
\url{https://canvas.education.lu.se/courses/34580/files/6061993?module_item_id=1441712}
}
\usage{
bloodpressure
}
\description{
A dataset used in \href{https://canvas.education.lu.se/courses/34580/files/6116314?module_item_id=1441704}{exercise 1}.
}
\keyword{datasets}
